refactor(create-expo-module): move environment variables to a shared `./utils/env` with docs
