+++ packages/create-expo-module/src/createExampleApp.ts
@@ -1,19 +1,17 @@
 import spawnAsync from '@expo/spawn-async';
-import getenv from 'getenv';
 import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 
-import { installDependencies } from './packageManager';
-import { PackageManagerName } from './resolvePackageManager';
-import { SubstitutionData } from './types';
+import { installDependencies, type PackageManagerName } from './packageManager';
+import type { SubstitutionData } from './types';
+import { env } from './utils/env';
 import { newStep } from './utils/ora';
 
 const debug = require('debug')('create-expo-module:createExampleApp') as typeof console.log;
 
 // These dependencies will be removed from the example app (`expo init` adds them)
 const DEPENDENCIES_TO_REMOVE = ['expo-status-bar', 'expo-splash-screen'];
-const EXPO_BETA = getenv.boolish('EXPO_BETA', false);
 
 /**
  * Initializes a new Expo project as an example app.
@@ -38,7 +36,7 @@ export async function createExampleApp(
   }
 
   await newStep('Initializing the example app', async (step) => {
-    const templateVersion = EXPO_BETA ? 'next' : 'latest';
+    const templateVersion = env.EXPO_BETA ? 'next' : 'latest';
     const template = `expo-template-blank-typescript@${templateVersion}`;
     debug(`Using example template: ${template}`);
     const command = createCommand(packageManager, exampleProjectSlug, template);
