+++ packages/create-expo-module/src/telemetry.ts
@@ -1,18 +1,13 @@
 import JsonFile from '@expo/json-file';
 import TelemetryClient from '@expo/rudder-sdk-node';
-import { boolish } from 'getenv';
 import crypto from 'node:crypto';
-import os, { homedir } from 'node:os';
+import os from 'node:os';
 import path from 'node:path';
 
-import { CommandOptions } from './types';
+import type { CommandOptions } from './types';
+import { env } from './utils/env';
 
 const packageJson = require('../package.json');
-
-/** If telemetry is disabled by the user */
-const EXPO_NO_TELEMETRY = boolish('EXPO_NO_TELEMETRY', false);
-/** If the tool is running in a sanboxed environment, either staging or local envs */
-const EXPO_SANDBOX = boolish('EXPO_STAGING', false) || boolish('EXPO_LOCAL', false);
 
 /** The telemetry client instance to use */
 let client: TelemetryClient | null = null;
@@ -22,7 +17,9 @@ let telemetryId: string | null = null;
 export function getTelemetryClient() {
   if (!client) {
     client = new TelemetryClient(
-      EXPO_SANDBOX ? '24TKICqYKilXM480mA7ktgVDdea' : '24TKR7CQAaGgIrLTgu3Fp4OdOkI', // expo unified,
+      env.EXPO_STAGING || env.EXPO_LOCAL
+        ? '24TKICqYKilXM480mA7ktgVDdea'
+        : '24TKR7CQAaGgIrLTgu3Fp4OdOkI', // expo unified,
       'https://cdp.expo.dev/v1/batch',
       {
         flushInterval: 300,
@@ -40,12 +37,12 @@ export function getTelemetryClient() {
 // The ~/.expo directory is used to store authentication sessions,
 // which are shared between EAS CLI and Expo CLI.
 function getExpoHomeDirectory() {
-  const home = homedir();
+  const home = os.homedir();
   if (process.env.__UNSAFE_EXPO_HOME_DIRECTORY) {
     return process.env.__UNSAFE_EXPO_HOME_DIRECTORY;
-  } else if (boolish('EXPO_STAGING', false)) {
+  } else if (env.EXPO_STAGING) {
     return path.join(home, '.expo-staging');
-  } else if (boolish('EXPO_LOCAL', false)) {
+  } else if (env.EXPO_LOCAL) {
     return path.join(home, '.expo-local');
   }
   return path.join(home, '.expo');
@@ -92,7 +89,7 @@ type Event = {
 };
 
 export async function logEventAsync(event: Event) {
-  if (EXPO_NO_TELEMETRY) {
+  if (env.EXPO_NO_TELEMETRY) {
     return;
   }
